requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
llamafactory>=0.9.3
